pyproject.toml
[project]
name = "spark-brainstorm-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "google-genai>=1.67.0",
]
